requirements.txt
attrs==19.3.0
cachetools==4.1.1
click==7.1.1
chardet==3.0.4
datadog==0.38.0
DateTime==4.3
DBUtils==1.3
decorator==4.4.2
google-auth>=1.14.1
google-api-core
google-api-python-client==1.10.0
google-cloud-billing-budgets==1.5.1
google-cloud-logging==1.15.0
google-cloud-resource-manager==0.30.2
google-cloud-storage==1.27.0
grpcio==1.42.0
markus[datadog]==2.2.0
protobuf==3.12.2
protobuf3-to-dict==0.1.5
PyMySQL==0.9.3
pytest==5.4.2
PyYAML==5.3.1
